refactor test setup
